# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtifactListing": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "artifacts"
        ],
        "type": "object"
      },
...
      "CheckpointListing": {
        "properties": {
          "checkpoints": {
            "description": "Newest first.",
            "items": {
              "$ref": "#/components/schemas/CheckpointEntry"
            },
            "type": "array"
          },
          "directory": {
            "description": "Absent when the run writes no checkpoints, which is not the same as an\nempty directory.",
            "type": [
              "string",
              "null"
            ]
          },
          "latest": {
            "description": "The id a resume would pick, i.e. what `fork_from` resolves to by default.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "checkpoints"
        ],
        "type": "object"
      },
      "CheckpointPatch": {
        "additionalProperties": false,
        "properties": {
          "every_steps": {
            "default": null,
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "mode": {
            "default": null,
            "description": "`steps` | `best_eval` | `steps_and_best_eval`, the same vocabulary the\nconfiguration document uses.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
...
      "CommandAccepted": {
        "description": "What every control route answers.\n\nOne shape for pause, resume, cancel, checkpoint and patch, because a client\nscripting them wants one thing to read: what the run is now, and when what it\nasked for lands.",
        "properties": {
          "applied": {
            "description": "The dotted paths a `PATCH` accepted, in a fixed order. Empty elsewhere.",
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "applies_at_iteration": {
            "description": "The iteration the request takes effect at - the next one, since a run is\nonly interruptible at its progress callback. Absent when the effect\nwas immediate, as for a run cancelled before it ever started.",
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "type": "object"
      },
...
      "DownloadLink": {
        "description": "`POST /v1/runs/{id}/artifacts/{name}/link`: a download a browser can\nnavigate to without a token.",
        "properties": {
          "expires_at": {
            "description": "Unix seconds after which the link is refused.",
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "href": {
            "description": "Same-origin path, query included. Good for this one artefact only.",
            "type": "string"
          }
        },
        "required": [
          "href",
          "expires_at"
        ],
        "type": "object"
      },
...
      "FieldError": {
        "description": "One field-level problem, located by a JSON Pointer (RFC 6901) into the\nrequest body.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "hint": {
            "description": "What to do about it, only when there is a short, concrete answer. A\nvague hint is worse than none.",
            "type": [
              "string",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "pointer": {
            "type": "string"
          }
        },
        "required": [
          "pointer",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "PatchRequest": {
        "additionalProperties": false,
        "description": "`PATCH /v1/runs/{id}` - the whitelist of the contract, and nothing else.\n\n`deny_unknown_fields` is what makes the list closed: `n_ctx`, `lora`,\n`algorithm` and every path are rejected by the deserializer before a handler\nruns, and the handler turns that into the 409 that points at `fork_from`.\nField names follow the *document* grammar (`training.lr`), the same one\n`params`, `provenance` and `GET /v1/defaults` speak.",
        "properties": {
          "checkpoint": {
            "default": null,
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/CheckpointPatch"
              }
            ]
          },
          "evaluation": {
            "default": null,
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EvaluationPatch"
              }
            ]
          },
          "training": {
            "default": null,
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TrainingPatch"
              }
            ]
          }
        },
        "type": "object"
      },
...
      "Problem": {
        "description": "Wire shape of a problem document (RFC 9457). Serialized by hand rather than\nderived from `ApiError` so the field order - and therefore the bytes - is\nfixed. Published as the `Problem` component: every failure of the API but\nthe OpenAI routes' has this body.",
        "properties": {
          "detail": {
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "meta": {
            "description": "The payload specific to the problem type, e.g. the dominant memory\nposts of `insufficient-memory`.",
            "type": [
              "object",
              "null"
            ]
          },
          "status": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "description": "The id every log line of this request carries.",
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "description": "`https://retrograd.dev/problems/<slug>`: the slug is what a client\nbranches on.",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "type": "object"
      },
...
      "Provenance": {
        "additionalProperties": {
          "$ref": "#/components/schemas/Origin"
        },
        "description": "The provenance map. `BTreeMap` so the serialized bytes are the same for two\nidentical resolutions (invariant 3).",
        "propertyNames": {
          "type": "string"
        },
        "type": "object"
      },
...
      "RunView": {
        "description": "`GET /v1/runs/{id}` and the `201` of `POST /v1/runs`.\n\nThe three heavy fields are carried as pre-rendered JSON. They are produced by\nserializing the typed resolution once, at creation, and then written to the\njournal and served verbatim - which is both cheaper than re-rendering and the\nonly way a run restored from disk answers the same bytes it answered while it\nwas alive.",
        "properties": {
          "algorithm": {
            "description": "The `run.algorithm` spelling.",
            "type": "string"
          },
          "created_at": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "effective_config": {
            "$ref": "#/components/schemas/ConfigDocument"
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "finished_at": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "holds_device": {
            "description": "Weights, KV caches and optimizer state stay allocated while a run is\npaused. True whenever this run is holding the device.",
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "model": {
            "description": "The base model's file name.",
            "type": "string"
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "objective": {
            "type": [
              "string",
              "null"
            ]
          },
          "observed": {
            "type": "boolean"
          },
          "plan": {
            "$ref": "#/components/schemas/PlanSummary"
          },
          "progress": {
            "$ref": "#/components/schemas/RunProgress"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "queue_position": {
            "format": "int32",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "started_at": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "id",
          "status",
          "algorithm",
          "model",
          "observed",
          "created_at",
          "progress",
          "holds_device",
          "effective_config",
          "provenance",
          "plan"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP control plane for LoRA training runs.",
    "title": "retrograd-server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/runs/{id}": {
      "delete": {
        "operationId": "delete_runs_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "deleted"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Forget a finished run"
      },
      "get": {
        "operationId": "get_runs_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "One run: state, progress, effective config, plan"
      },
      "patch": {
        "operationId": "patch_runs_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAccepted"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Adjust a schedule on a live run"
      }
    },
    "/v1/runs/{id}/artifacts": {
      "get": {
        "operationId": "get_runs_id_artifacts",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactListing"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "What this run wrote"
      }
    },
    "/v1/runs/{id}/artifacts/{name}": {
      "get": {
        "operationId": "get_runs_id_artifacts_name",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Signature of a link from POST …/link; with `exp`, stands in for the bearer token.",
            "in": "query",
            "name": "sig",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Expiry of that link, unix seconds.",
            "in": "query",
            "name": "exp",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Download one artifact"
      }
    },
    "/v1/runs/{id}/artifacts/{name}/link": {
      "post": {
        "operationId": "post_runs_id_artifacts_name_link",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadLink"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "A signed, one-minute link to download one artifact without a token"
      }
    },
...
    "/v1/runs/{id}/checkpoints": {
      "get": {
        "operationId": "get_runs_id_checkpoints",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckpointListing"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Checkpoints on disk"
      },
      "post": {
        "operationId": "post_runs_id_checkpoints",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAccepted"
                }
              }
            },
            "description": "ok"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "a problem document (RFC 9457)"
          }
        },
        "summary": "Ask for a checkpoint at the next boundary"
      }
    },
...
  }
}